pyproject.toml
[project]
name = "docmind"
version = "1.0.0"
description = "Document Q&A agent with RAG (ChromaDB + Groq) and an MCP server for AI agent integration."
readme = "README.md"
requires-python = ">=3.11, <3.13"
license = {text = "MIT"}
dependencies = [
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-groq==0.2.1",
    "chromadb==0.5.18",
    "sentence-transformers==3.3.1",
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "python-dotenv==1.0.1",
    "pypdf==5.1.0",
    "httpx==0.27.2",
    "gradio==4.44.1",
    "mcp==1.0.0",
    "pydantic<2.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "mcp_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
